Implement event assertions against UnexpectedEvents.
Rework the internals so the first|second|third checks are assertions. Use
them for the existing assertions as well as exposing them outwardly.

Latch onto the fact a series of captured events are stored with their own
identifiable type to allow the event comparisons to be come assertions.
This should complete the infrastructure required to allow fully splitting
event capture from later asserting what was present.